refactor : merge ressource_gift and information_gift_log into one file
Both flows share the same TestConfig scenario and controller-id lookup
helpers; combining them saves one scenario reload (~30s) and ~75 lines
of duplicated helpers/fixture boilerplate.
