+++ tests/test_gift_mechanics_e2e.py
@@ -1,17 +1,11 @@
-"""Playwright E2E tests for Issue #10 — ressource gift mechanic.
+"""Playwright E2E tests for the two gift mechanics — ressource transfer
+(Issue #10) and information gifts with the per-gift log (Issue #61).
 
-Tests rely on TestConfig CSV defaults (Alpha 100 Gold, Beta 80 Gold, etc.)
-and verify behaviour by reading the rendered HTML of the Ressources page.
-
-No direct DB writes. The IDs needed to build form payloads are scraped
-from the page's own <option value="..."> attributes so the tests can
-run on the demo environment with UI_ONLY=1.
-
-Assertions are delta-based — they compare pre/post amounts read from
-the UI, so they don't break if a prior test nudged the baseline.
+Both flows share the same TestConfig scenario and the same controller-id
+lookup helpers, so they live in one module to share the scenario load.
 
 Run:
-    python3 -m pytest tests/test_ressource_gift_e2e.py -v
+    python3 -m pytest tests/test_gift_mechanics_e2e.py -v
 """
 import re
 
@@ -19,7 +13,7 @@ import pytest
 
 from conftest import PHP_BASE_URL, ensure_gm_login
 from helpers import (
-    DB_AVAILABLE, load_minimal_data, safe_goto,
+    DB_AVAILABLE, load_minimal_data, load_scenario_via_admin, safe_goto,
     register_php_error_listener, assert_no_collected_php_errors,
 )
 
@@ -33,21 +27,17 @@ def base_url():
 def load_test_config(browser):
     if DB_AVAILABLE:
         load_minimal_data()
-    context = browser.new_context()
-    page = context.new_page()
-    ensure_gm_login(page, PHP_BASE_URL)
-    safe_goto(page, f"{PHP_BASE_URL}/base/admin.php")
-    page.wait_for_load_state("networkidle")
-    page.locator("select[name='config_name']").select_option("TestConfig")
-    page.locator("input[name='submit'][value='Submit']").click()
-    page.wait_for_timeout(5000)
-    page.wait_for_load_state("load", timeout=90000)
-    context.close()
+    load_scenario_via_admin(browser, PHP_BASE_URL, "TestConfig")
     yield
 
 
+# ---------------------------------------------------------------------------
+# Shared lookup + view helpers
+# ---------------------------------------------------------------------------
+
+
 def _resolve_controller_id_via_ui(page, lastname):
-    """Read a controller_id from accueil's universal select (shows all controllers)."""
+    """Read a controller_id from accueil's universal select (shows all)."""
     safe_goto(page, f"{PHP_BASE_URL}/base/accueil.php")
     page.wait_for_load_state("load")
     return page.locator(
@@ -59,10 +49,31 @@ def _set_active_via_ui(page, lastname):
     value = _resolve_controller_id_via_ui(page, lastname)
     safe_goto(page, f"{PHP_BASE_URL}/base/accueil.php?controller_id={value}")
     page.wait_for_load_state("load")
+    return value
 
 
-def _view(page):
+def _resolve_location_id_via_ui(page, location_name):
+    safe_goto(page, f"{PHP_BASE_URL}/zones/management_locations.php")
+    page.wait_for_load_state("load")
+    html = page.content()
+    m = re.search(
+        rf'<h3>[^<]*{re.escape(location_name)}[^<]*\(discovery[^<]+</h3>'
+        rf'.*?name="toggle_destruction"\s+value="(\d+)"',
+        html, re.DOTALL,
+    )
+    if not m:
+        raise AssertionError(f"location_id for {location_name!r} not found")
+    return m.group(1)
+
+
+def _ressources_view(page):
     safe_goto(page, f"{PHP_BASE_URL}/ressources/view.php")
+    page.wait_for_load_state("load")
+    return page.content()
+
+
+def _faction_view(page):
+    safe_goto(page, f"{PHP_BASE_URL}/controllers/action.php")
     page.wait_for_load_state("load")
     return page.content()
 
@@ -74,6 +85,23 @@ def _scrape_amount(content, ressource_name):
     return int(m.group(1)) if m else None
 
 
+def _seed_ckl_admin(page, recipient_id, location_id):
+    """Admin path writes CKL row without an information-gift-log entry —
+    sets up known-location state for the player-path gift under test."""
+    safe_goto(
+        page,
+        f"{PHP_BASE_URL}/controllers/management.php"
+        f"?giftInformationLocation=1&target_controller_id={recipient_id}"
+        f"&location_id={location_id}"
+    )
+    page.wait_for_load_state("load")
+
+
+# ---------------------------------------------------------------------------
+# Ressource gift (Issue #10)
+# ---------------------------------------------------------------------------
+
+
 class TestRessourcesPageRenders:
     """Page loads with TestConfig amounts and gift form visible."""
 
@@ -83,12 +111,10 @@ class TestRessourcesPageRenders:
         page = ctx.new_page()
         register_php_error_listener(page)
         ensure_gm_login(page, PHP_BASE_URL)
-
         _set_active_via_ui(page, "Alpha")
-        content = _view(page)
+        content = _ressources_view(page)
         assert_no_collected_php_errors(page)
         ctx.close()
-
         type(self)._content = content
         yield
 
@@ -106,7 +132,7 @@ class TestRessourcesPageRenders:
         assert _scrape_amount(self._content, "Gold") is not None
 
 
-class TestGiftHappyPath:
+class TestRessourceGiftHappyPath:
     """Submitting the gift form via UI shifts the giver and recipient amounts."""
 
     _gift_amount = 10
@@ -119,10 +145,9 @@ class TestGiftHappyPath:
         ensure_gm_login(page, PHP_BASE_URL)
 
         _set_active_via_ui(page, "Alpha")
-        pre_alpha = _scrape_amount(_view(page), "Gold")
-
+        pre_alpha = _scrape_amount(_ressources_view(page), "Gold")
         _set_active_via_ui(page, "Beta")
-        pre_beta = _scrape_amount(_view(page), "Gold")
+        pre_beta = _scrape_amount(_ressources_view(page), "Gold")
 
         _set_active_via_ui(page, "Alpha")
         safe_goto(page, f"{PHP_BASE_URL}/ressources/view.php")
@@ -142,12 +167,11 @@ class TestGiftHappyPath:
         post_alpha = _scrape_amount(post_alpha_content, "Gold")
 
         _set_active_via_ui(page, "Beta")
-        post_beta_content = _view(page)
+        post_beta_content = _ressources_view(page)
         post_beta = _scrape_amount(post_beta_content, "Gold")
 
         assert_no_collected_php_errors(page)
         ctx.close()
-
         type(self)._pre_alpha = pre_alpha
         type(self)._post_alpha = post_alpha
         type(self)._pre_beta = pre_beta
@@ -172,7 +196,7 @@ class TestGiftHappyPath:
         assert "Gold" in self._post_beta_content
 
 
-class TestGiftValidationRejects:
+class TestRessourceGiftValidationRejects:
     """Self-target, over-stock, and zero-amount POSTs leave both amounts unchanged."""
 
     @pytest.fixture(scope="class", autouse=True)
@@ -186,31 +210,30 @@ class TestGiftValidationRejects:
         beta_value = _resolve_controller_id_via_ui(page, "Beta")
 
         _set_active_via_ui(page, "Alpha")
-        pre_alpha_content = _view(page)
+        pre_alpha_content = _ressources_view(page)
         pre_alpha = _scrape_amount(pre_alpha_content, "Gold")
         gold_value = page.locator(
             "select[name='ressource_id'] option:has-text('Gold')"
         ).first.get_attribute("value")
 
         _set_active_via_ui(page, "Beta")
-        pre_beta = _scrape_amount(_view(page), "Gold")
+        pre_beta = _scrape_amount(_ressources_view(page), "Gold")
 
         _set_active_via_ui(page, "Alpha")
         rejections = [
             {"ressource_id": gold_value, "target_controller_id": beta_value,  "amount": "999999", "giftRessource": "Donner"},
-            {"ressource_id": gold_value, "target_controller_id": alpha_value, "amount": "5",     "giftRessource": "Donner"},
+            {"ressource_id": gold_value, "target_controller_id": alpha_value, "amount": "5",      "giftRessource": "Donner"},
             {"ressource_id": gold_value, "target_controller_id": beta_value,  "amount": "0",      "giftRessource": "Donner"},
         ]
         for form in rejections:
             page.request.post(f"{PHP_BASE_URL}/ressources/action.php", form=form)
 
-        post_alpha = _scrape_amount(_view(page), "Gold")
+        post_alpha = _scrape_amount(_ressources_view(page), "Gold")
         _set_active_via_ui(page, "Beta")
-        post_beta = _scrape_amount(_view(page), "Gold")
+        post_beta = _scrape_amount(_ressources_view(page), "Gold")
 
         assert_no_collected_php_errors(page)
         ctx.close()
-
         type(self)._pre_alpha = pre_alpha
         type(self)._post_alpha = post_alpha
         type(self)._pre_beta = pre_beta
@@ -222,3 +245,92 @@ class TestGiftValidationRejects:
 
     def test_recipient_unchanged(self):
         assert self._post_beta == self._pre_beta
+
+
+# ---------------------------------------------------------------------------
+# Information gift log (Issue #61)
+# ---------------------------------------------------------------------------
+
+
+class TestInformationPanelPresent:
+    """Faction page renders the 'Informations reçues' panel, empty by default."""
+
+    @pytest.fixture(scope="class", autouse=True)
+    def panel_state(self, browser):
+        ctx = browser.new_context()
+        page = ctx.new_page()
+        register_php_error_listener(page)
+        ensure_gm_login(page, PHP_BASE_URL)
+        _set_active_via_ui(page, "Alpha")
+        content = _faction_view(page)
+        assert_no_collected_php_errors(page)
+        ctx.close()
+        type(self)._content = content
+        yield
+
+    def test_panel_present(self):
+        assert "Informations reçues" in self._content
+
+
+class TestInformationGiftLocationLogged:
+    """Player-path giftInformationLocation must write to information_gift_logs.
+    The recipient's Faction page should then show an entry in the panel."""
+
+    @pytest.fixture(scope="class", autouse=True)
+    def gift_state(self, browser):
+        ctx = browser.new_context()
+        page = ctx.new_page()
+        register_php_error_listener(page)
+        ensure_gm_login(page, PHP_BASE_URL)
+
+        alpha_id = _resolve_controller_id_via_ui(page, "Alpha")
+        beta_id = _resolve_controller_id_via_ui(page, "Beta")
+        echo_base_id = _resolve_location_id_via_ui(page, "Echo-Base")
+
+        _seed_ckl_admin(page, alpha_id, echo_base_id)
+
+        _set_active_via_ui(page, "Alpha")
+        page.request.get(
+            f"{PHP_BASE_URL}/controllers/action.php"
+            f"?giftInformationLocation=1"
+            f"&controller_id={alpha_id}"
+            f"&target_controller_id={beta_id}"
+            f"&location_id={echo_base_id}"
+        )
+
+        _set_active_via_ui(page, "Beta")
+        beta_content = _faction_view(page)
+        assert_no_collected_php_errors(page)
+        ctx.close()
+        type(self)._beta_content = beta_content
+        yield
+
+    def test_recipient_sees_giver_lastname(self):
+        assert "Alpha" in self._beta_content
+
+    def test_recipient_sees_target_location_name(self):
+        assert "Echo-Base" in self._beta_content
+
+    def test_recipient_sees_transmis_phrasing(self):
+        assert "vous a transmis" in self._beta_content
+
+
+class TestInformationAdminTransactionsTable:
+    """controllers/management.php gains an 'Information Transactions' admin section."""
+
+    @pytest.fixture(scope="class", autouse=True)
+    def admin_state(self, browser):
+        ctx = browser.new_context()
+        page = ctx.new_page()
+        register_php_error_listener(page)
+        ensure_gm_login(page, PHP_BASE_URL)
+        safe_goto(page, f"{PHP_BASE_URL}/controllers/management.php")
+        page.wait_for_load_state("load")
+        admin_content = page.content()
+        assert_no_collected_php_errors(page)
+        ctx.close()
+        type(self)._admin_content = admin_content
+        yield
+
+    def test_admin_section_present(self):
+        assert "Information Transactions" in self._admin_content
